Switched to reactive stack

+++ src/main/kotlin/pl/komjago/ticketapp/controllers/booking/dto/BookedSeatInfo.kt
@@ -1,6 +1,6 @@
 package pl.komjago.ticketapp.controllers.booking.dto
 
 data class BookedSeatInfo(
-        val ticketTypeId: Long,
-        val seatId: Long
+        val ticketTypeId: Int,
+        val seatId: Int
 )
